more fixes and code quality
- Action PIF now checks if module exists before processing.
- Inlined a few print error checks.
- Other fixes...

+++ magisk_module_files/action.sh
@@ -48,6 +48,11 @@ PlayIntegrityFix() {
   PIF_MODULE_DIR="/data/adb/modules/playintegrityfix"
   PIF_DIRS="$PIF_MODULE_DIR/pif.json"
 
+  # Check if PIF_MODULE_DIR is a directory and module.prop exists and is not empty
+  if [[ ! (-d "$PIF_MODULE_DIR" && -s "$PIF_MODULE_DIR/module.prop") ]]; then
+    abort "PlayIntegrityFix module is missing or is not accessible, Skipping !"
+  fi
+
   # Check whether we are about to build pif.json for PIF Fork or not
   if grep -q "Fork" "$PIF_MODULE_DIR/module.prop"; then
     ui_print " ! Detected a Fork version of PlayIntegrityFix, Please install the official version !"
